Move the requestrecord infrastructure to modules/web/routing
